requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
psycopg2
python-dotenv
Werkzeug